+++ clients/fides-js/src/components/Overlay.tsx
@@ -27,6 +27,7 @@ import type { I18n } from "../lib/i18n";
 import ConsentModal from "./ConsentModal";
 import ConsentContent from "./ConsentContent";
 import "./fides.css";
+import { blockPageScrolling, unblockPageScrolling } from "../lib/ui-utils";
 
 interface RenderBannerProps {
   isOpen: boolean;
@@ -51,6 +52,7 @@ interface Props {
   renderModalContent: () => VNode;
   renderModalFooter: (props: RenderModalFooter) => VNode;
   onVendorPageClick?: () => void;
+  isUiBlocking: boolean;
 }
 
 const Overlay: FunctionComponent<Props> = ({
@@ -65,12 +67,25 @@ const Overlay: FunctionComponent<Props> = ({
   renderModalContent,
   renderModalFooter,
   onVendorPageClick,
+  isUiBlocking,
 }) => {
   const delayBannerMilliseconds = 100;
   const delayModalLinkMilliseconds = 200;
   const hasMounted = useHasMounted();
   const [bannerIsOpen, setBannerIsOpen] = useState(false);
   const modalLinkRef = useRef<HTMLElement | null>(null);
+
+  useEffect(() => {
+    if (isUiBlocking && bannerIsOpen) {
+      blockPageScrolling();
+    } else {
+      unblockPageScrolling();
+    }
+
+    return () => {
+      unblockPageScrolling();
+    };
+  }, [isUiBlocking, bannerIsOpen]);
 
   const dispatchCloseEvent = useCallback(
     ({ saved = false }: { saved?: boolean }) => {
@@ -177,9 +192,10 @@ const Overlay: FunctionComponent<Props> = ({
 
   return (
     <div>
-      {showBanner && bannerIsOpen && window.Fides.options.preventDismissal && (
+      {showBanner && bannerIsOpen && isUiBlocking && (
         <div className="fides-modal-overlay" />
       )}
+
       {showBanner
         ? renderBanner({
             isOpen: bannerIsOpen,